Separated business logic to FirstVCViewModel and created a model for flight data.  Corrected print buttons to fire.
